app/requirements.txt
requests
psycopg2-binary
# boto3